Physics

+++ inc/libftmlx/math/ftmlx_mat.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ftmlx_mat.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kmendes <kmendes@student.42lausanne.ch>    +#+  +:+       +#+        */
+/*   By: kmendes <kmendes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/31 03:11:28 by kmendes           #+#    #+#             */
-/*   Updated: 2022/06/25 02:50:41 by kmendes          ###   ########.fr       */
+/*   Updated: 2022/11/03 16:45:42 by kmendes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 typedef struct s_mat4
 {
-	float	m[4][4];
+	double	m[4][4];
 }	t_mat4;
 
 t_fvec4	mat_mult_vec(t_mat4 m, t_fvec4 v);
